+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 FEATURES:
 - Updated github actions
 - Added release automation
+- Added CHANGELOG.md
+- Updated minimum terraform version to resolve this [issue](https://github.com/hashicorp/terraform/issues/32200)
+- Updated minimum azurerm version to [3.73.0](https://github.com/hashicorp/terraform-provider-azurerm/releases/tag/v3.73.0)
 
 BUG FIXES:
 - Added null value validation
+- Updated go mod github.com/Azure/go-autorest/autorest/adal@latest